requirements.txt
torch
torchvision
diffusers
transformers
accelerate
tqdm
pillow
gradio